packages/snaps-cli/src/webpack/compiler.ts
import { webpack } from 'webpack';

import type { ProcessedWebpackConfig } from '../config';
import type { WebpackOptions } from './config';
import { getDefaultConfiguration } from './config';

/**
 * Get a Webpack compiler for the given config.
 *
 * @param config - The config object.
 * @param options - The Webpack options.
 * @returns The Webpack compiler.
 */
export async function getCompiler(
  config: ProcessedWebpackConfig,
  options?: WebpackOptions,
) {
  const baseWebpackConfig = await getDefaultConfiguration(config, options);
  const webpackConfig =
    config.customizeWebpackConfig?.(baseWebpackConfig) ?? baseWebpackConfig;

  return webpack(webpackConfig);
}
